Switch to using a service account for authentication
Before this, I was using OAuth2 credentials linked to my (human-user) Google
account, which meant that I had to grant it permission to access *all* of my
content in Google Sheets, which is obviously far from ideal.

With this approach, we can give the service account edit access to the
relevant spreadsheet(s) using its email address, and give it no other
permissions. This means it's safer to use, especially in a deployment
pipeline. There's also no OAuth user flow to authorise it in the browser,
which is more appropriate for this use case.

I created the service account for development with no roles, so I don't think
it has access to anything in the Google APIs project. Service accounts may
work a bit differently in the context of a GSuite domain, but I'm hoping we
can still keep the service account permissions locked down to just the
spreadsheets we need to manage.

https://developers.google.com/identity/protocols/OAuth2ServiceAccount