Combine risk register criteria into one more general question
Risk registers (formally defined) aren't an appropriate way of recording
security risks for every service, and by including references to it we risk
creating incentives for teams to use inappropriate methods and make it harder
to assess our real situation accurately.

We introduced these criteria in an attempt to pin down what having a good
understanding of the security situation of an application meant, but we think
it's better to go back to that more general question at this stage. We'll
continue to work with the security team to develop a more precise and useful
way of assessing this point.

This change means removing a column, so everything to the right shifts over by
one column. This kind of change would be easier if we kept a mapping of column
name to letter, and then used that in all references to columns, but that
could also get a bit difficult to read. Worth seeing how that would look,
though.