Update conditional format rule index numbers
After moving the tech individual criteria rules above the atrophy ones. It
would be nice to not have to worry about this in general. Earlier indexed
rules override later ones, so we need red rules to come before amber before
green for rules which affect the same cells - but the ordering of rules which
affect different cells doesn't have any impact.