Protect the header rows from updates
Teams are responsible for keeping the data up to date, so many people will
need to have access to edit the spreadsheet. But those people shouldn't be
able to change the structure of the spreadsheet.

This change prevents users from editing the two header rows, except for those
specified as editors.