Switch to specifying sheet by title, not ID
The sheet id isn't the same as the sheet index (although I'm sure I tested
that when I first started working on this... oh well) and users aren't going
to be able to provide the sheet id easily. The sheet index is also liable to
change as tabs are reordered, so the title is the safest way to specify which
sheet in a spreadsheet should be managed by this script.