Check the number of columns before modifying the sheet
And exit if it's not the expected number.

It's going to be too tricky to handle schema migrations in this script when we
add, remove or merge criteria, and not worth the effort for the frequency of
changes we expect to make. Migrating existing data between columns is best
done manually in the cases we know about at the moment, so this script also
won't try to handle that.

This check and error message will remind us to make those changes manually
instead when necessary.