Start a changelog
We're about to start making some breaking changes which will require manual
edits to spreadsheets before running later versions of the script. This
changelog will document how to make those manual edits needed around each
breaking change, which so far arise from us iterating on the criteria and
merging some.

This script will not handle schema migrations because it's not worth the
effort at this stage of setting up a deployment mechanism that would handle
them reliably. It won't handle data migrations around schema changes because
that would mean trying to predict how other people's data should be merged,
and I'm not going to attempt that.